Refactor and add option to build Velox test utilities
