+++ velox/dwio/dwrf/CMakeLists.txt
@@ -17,6 +17,8 @@ add_subdirectory(proto)
 add_subdirectory(reader)
 if(${VELOX_BUILD_TESTING})
   add_subdirectory(test)
+elseif(${VELOX_BUILD_TEST_UTILS})
+  add_subdirectory(test/utils)
 endif()
 add_subdirectory(utils)
 add_subdirectory(writer)
