+++ velox/exec/tests/utils/CMakeLists.txt
@@ -12,7 +12,28 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_library(velox_exec_test_util FunctionUtils.cpp TempFilePath.cpp
-                                 TempDirectoryPath.cpp)
-target_link_libraries(velox_exec_test_util velox_core velox_exception
-                      velox_expression)
+add_library(
+  velox_exec_test_util
+  FunctionUtils.cpp
+  TempFilePath.cpp
+  TempDirectoryPath.cpp
+  PlanBuilder.cpp
+  QueryAssertions.cpp
+  Cursor.cpp
+  OperatorTestBase.cpp
+  HiveConnectorTestBase.cpp)
+
+target_link_libraries(
+  velox_exec_test_util
+  duckdb
+  velox_core
+  velox_exception
+  velox_expression
+  velox_parse_parser
+  velox_duckdb_conversion
+  velox_dwio_common
+  velox_dwio_type_fbhive
+  velox_dwio_dwrf_reader
+  velox_dwio_dwrf_writer
+  velox_dwrf_test_utils
+  velox_presto_serializer)
