+++ velox/exec/tests/utils/HiveConnectorTestBase.h
@@ -16,7 +16,7 @@
 #pragma once
 #include "velox/connectors/hive/HiveConnector.h"
 #include "velox/exec/Operator.h"
-#include "velox/exec/tests/OperatorTestBase.h"
+#include "velox/exec/tests/utils/OperatorTestBase.h"
 #include "velox/exec/tests/utils/TempFilePath.h"
 #include "velox/type/tests/SubfieldFiltersBuilder.h"
