+++ velox/exec/tests/utils/OperatorTestBase.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "velox/exec/tests/OperatorTestBase.h"
+
 #include <velox/parse/Expressions.h>
 #include <velox/parse/ExpressionsParser.h>
 #include "velox/common/caching/AsyncDataCache.h"
@@ -21,6 +21,7 @@
 #include "velox/exec/Exchange.h"
 #include "velox/exec/PartitionedOutputBufferManager.h"
 #include "velox/exec/tests/utils/FunctionUtils.h"
+#include "velox/exec/tests/utils/OperatorTestBase.h"
 #include "velox/functions/prestosql/SimpleFunctions.h"
 #include "velox/functions/prestosql/VectorFunctions.h"
 #include "velox/serializers/PrestoSerializer.h"
