+++ velox/experimental/codegen/tests/CMakeLists.txt
@@ -20,7 +20,6 @@ target_link_libraries(
   velox_core
   velox_exec
   velox_experimental_codegen
-  velox_exec_test_lib
   velox_exec_test_util
   velox_functions_prestosql
   velox_codegen_utils_resource_path)
