+++ velox/experimental/codegen/tests/CodegenTestBase.h
@@ -23,9 +23,9 @@
 #include "velox/core/PlanNode.h"
 #include "velox/dwio/dwrf/test/utils/BatchMaker.h"
 #include "velox/exec/Operator.h"
-#include "velox/exec/tests/Cursor.h"
-#include "velox/exec/tests/PlanBuilder.h"
+#include "velox/exec/tests/utils/Cursor.h"
 #include "velox/exec/tests/utils/FunctionUtils.h"
+#include "velox/exec/tests/utils/PlanBuilder.h"
 #include "velox/experimental/codegen/CodegenCompiledExpressionTransform.h"
 #include "velox/experimental/codegen/CodegenLogger.h"
 #include "velox/experimental/codegen/code_generator/ExprCodeGenerator.h"
