+++ velox/experimental/codegen/tests/TARGETS_disabled
@@ -16,7 +16,6 @@ cpp_library(
         "//velox/core:velox_core",
         "//velox/dwio/dwrf/test/utils:lib_test_batch_maker",
         "//velox/exec:velox_exec_lib",
-        "//velox/exec/tests:velox_exec_test_lib",
         "//velox/exec/tests/utils:velox_exec_test_util",
         "//velox/experimental/codegen:velox_experimental_codegen_header",
         "//velox/experimental/codegen:velox_experimental_codegen_impl",
