+++ velox/experimental/codegen/vector_function/tests/CMakeLists.txt
@@ -21,7 +21,6 @@ target_link_libraries(
   velox_codegen_vector_function_test
   velox_codegen_ast
   velox_codegen_external_process
-  velox_exec_test_lib
   velox_exec
   velox_dwio_dwrf_reader
   velox_dwio_dwrf_writer
