+++ velox/functions/lib/tests/CMakeLists.txt
@@ -22,7 +22,6 @@ target_link_libraries(
   velox_functions_lib_test
   velox_functions_lib
   velox_functions_test_lib
-  velox_exec_test_lib
   velox_exec_test_util
   velox_expression
   ${GTEST_BOTH_LIBRARIES}
