+++ velox/functions/prestosql/aggregates/tests/CMakeLists.txt
@@ -14,7 +14,7 @@
 
 add_library(velox_aggregates_test_lib AggregationTestBase.cpp)
 
-target_link_libraries(velox_aggregates_test_lib velox_exec_test_lib)
+target_link_libraries(velox_aggregates_test_lib velox_exec_test_util)
 
 add_executable(
   velox_aggregates_test
@@ -41,7 +41,6 @@ add_test(
 
 target_link_libraries(
   velox_aggregates_test
-  velox_exec_test_lib
   velox_functions_test_lib
   velox_exec
   velox_aggregates
